+++ src/widgets/Clickable.tsx
@@ -1,0 +1,15 @@
+import * as React from "react";
+
+type ChildProps = {
+  className?: string;
+};
+
+type PropsType = {
+  renderChild: (props: ChildProps) => React.ReactElement;
+};
+
+export default function Clickable(props: PropsType) {
+  const { renderChild } = props;
+
+  return renderChild({});
+}
